Fix price handling to avoid rounding errors and add order update tests
Fixes #99

+++ src/main/java/com/vaadin/template/orders/ui/view/orders/OrderEditPresenter.java
@@ -90,8 +90,8 @@ public class OrderEditPresenter {
 	}
 
 	private void updateTotalSum() {
-		double sum = view.getOrder().getItems().stream().filter(item -> item.getProduct() != null)
-				.collect(Collectors.summingDouble(item -> item.getProduct().getPrice() * item.getQuantity()));
+		int sum = view.getOrder().getItems().stream().filter(item -> item.getProduct() != null)
+				.collect(Collectors.summingInt(item -> item.getProduct().getPrice() * item.getQuantity()));
 		view.setSum(sum);
 	}
